♻️ refactor : 구입비용 및 구입개수 기능 작성
createPurchaseAmount 메서드를
구입비용과 그에따른 구입개수 반환하도록 변경
