+++ src/lotto/index.js
@@ -1,0 +1,2 @@
+export { default as PurchaseAmount } from './PurchaseAmount.js';
+export { default as Lotto } from './Lotto.js';
