+++ src/utils/create-purchase-amount.js
@@ -1,9 +1,12 @@
-import { PurchaseAmount } from '../purchase-amount/index.js';
+import { PurchaseAmount } from '../lotto/index.js';
 import { inputMethod } from './index.js';
-import { inputConstants } from '../constants/index.js';
+import { uiConstants, magicNumber } from '../constants/index.js';
 
 export default async function createPurchaseAmount() {
-  const inputAmount = await inputMethod(inputConstants.LOTTO_BUY_MONEY);
+  const inputAmount = await inputMethod(uiConstants.LOTTO_BUY_MONEY_MESSAGE);
   const purchaseAmount = new PurchaseAmount(inputAmount);
-  return purchaseAmount;
+  return [
+    purchaseAmount.getPurchaseAmount(),
+    purchaseAmount.getPurchaseAmount() / magicNumber.UNIT,
+  ];
 }
