+++ src/utils/index.js
@@ -1,3 +1,3 @@
-export { default as inputMethod } from './input-method.js';
+export { inputMethod, printMethod } from './ui-method.js';
 export { default as createPurchaseAmount } from './create-purchase-amount.js';
 export { default as validatePurchaseAmount } from './validate-purchase-amount.js';
