+++ src/utils/ui-method.js
@@ -1,0 +1,11 @@
+import { MissionUtils } from '@woowacourse/mission-utils';
+import { uiConstants } from '../constants/index.js';
+
+export async function inputMethod(inputConst) {
+  const input = await MissionUtils.Console.readLineAsync(inputConst);
+  return input;
+}
+
+export function printMethod(printConst) {
+  const print = MissionUtils.Console.print(printConst);
+}
